requirements.txt
traja
alphashape
pandas
scipy
pillow
opencv-python
moviepy
imageio
numpy
shapely
concave_hull
